pyproject.toml
[tool.poetry]
name = "notice-board"
version = "0.1.0"
description = ""
authors = ["antos07 <trotsenko.ant07@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2"
psycopg = {extras = ["binary"], version = "^3.1.8"}
djangorestframework = "^3.14.0"
markdown = "^3.4.3"
drf-spectacular = "^0.26.2"
drf-nested-routers = "^0.93.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
